# docs/swagger.yaml
definitions:
  Error:
    properties:
      detail:
        type: string
    type: object
...
  Msg:
    properties:
      msg:
        type: string
    type: object
  Post:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    required:
    - content
    type: object
  PostCreate:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  PostUpdate:
    properties:
      content:
        type: string
    type: object
...
  User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      isSuperuser:
        type: boolean
      password:
        minLength: 8
        type: string
      updatedAt:
        type: string
    required:
    - email
    - password
    type: object
  UserCreate:
    properties:
      email:
        type: string
      fullName:
        type: string
      isActive:
        type: boolean
      isSuperuser:
        type: boolean
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  UserUpdate:
    properties:
      fullName:
        type: string
      isActive:
        type: boolean
      isSuperuser:
        type: boolean
      password:
        minLength: 8
        type: string
    type: object
  ValidationError:
    properties:
      detail:
        additionalProperties: true
        type: object
    type: object
info:
  contact: {}
  title: Start
  version: 0.1.0
paths:
...
  /api/v1/posts:
    get:
      consumes:
      - application/json
      description: Read posts
      parameters:
      - description: User Id
        in: query
        name: userId
        type: string
      - description: Search
        in: query
        name: search
        type: string
      - default: 0
        description: Skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Post'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Read Posts
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Create post
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/PostCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Post'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Create Post
      tags:
      - Posts
  /api/v1/posts/{post_id}:
    delete:
      consumes:
      - application/json
      description: Delete post
      parameters:
      - description: Post Id
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Msg'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Delete Post
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: Read post
      parameters:
      - description: Post Id
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Post'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Read Post
      tags:
      - Posts
    patch:
      consumes:
      - application/json
      description: Update post
      parameters:
      - description: Post Id
        in: path
        name: post_id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/PostUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Post'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Update Post
      tags:
      - Posts
  /api/v1/posts/home:
    get:
      consumes:
      - application/json
      description: Read home posts
      parameters:
      - description: Search
        in: query
        name: search
        type: string
      - default: 0
        description: Skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Post'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Read Home Posts
      tags:
      - Posts
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Read users
      parameters:
      - description: Follower Id
        in: query
        name: followerId
        type: string
      - description: Followed Id
        in: query
        name: followedId
        type: string
      - description: Search
        in: query
        name: search
        type: string
      - default: 0
        description: Skip
        in: query
        name: skip
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/User'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Read Users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/User'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      summary: Create User
      tags:
      - Users
  /api/v1/users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Msg'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Read user
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Read User
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"